#7291667 No more byte size handling in Type.
Smp::IType no longer have any accessor on the type byte size, then there
is nothing to do with such size and requesting a byte size on type
creation is useles since at the end, nothing can be done through the
standardized API with the provided size. Other means have to be setup
locally when a byte size is needed.